Download wind data from the Climate Forecast System Reanalysis (CFSR)

This function downloads hourly CFSR wind data from the NCAR THREDDS server, for one or more historic dates, within a spatial bounding box. Data for all factorial combinations of 'years', 'months', and 'days' are downloaded and combined.

Usage

cfsr_dl(
  years = 1979,
  months = 1,
  days = 1,
  xlim = c(260, 270),
  ylim = c(40, 50)
)

Arguments

years

Integer vector representing the historic year(s) for which data are to be accessed, in the range 1979-2010.

months

Integer vector representing the month(s) for which data are to be accessed, in the range 1-12.

days

Integer vector representing day(s) for which data is to be requested, in the range 0-31.

xlim

Vector of length 2, giving longitudinal limits of the data to download, in the range 0 to 360.

ylim

Vector of length 2, giving latitudinal limits of the data to download, in the range -90 to 90.

Value

A list containing u and v components of the wind field, each as a SpatRaster with a layer for each hour of each requested date.